+++ assets/gulp/tasks/unitTestRunner/karma/unit-runner.js
@@ -147,7 +147,17 @@ gulp.task("unit-run-test", async () => {
 
     if (options.browser) {
         karmaConf.singleRun = false;
-        karmaConf.browsers = options.browser.split(",");
+        karmaConf.browsers = [];
+        const overrideBrowsers = options.browser.split(",");
+        const allOptions = ["Chrome", "ChromeHeadless", "Edge", "Firefox", "IE", "PhantomJS", "Safari"];
+
+        overrideBrowsers.forEach(browser => {
+            const bLower = browser.toLowerCase();
+            const found = allOptions.find(option => option.toLowerCase() === bLower);
+            if (found) {
+                karmaConf.browsers.push(found);
+            }
+        });
     }
 
     if (options.watch) {